+++ src/components/common/Checkbox.tsx
@@ -1,26 +1,50 @@
 import React from "react";
+import classNames from "classnames";
 
 import styles from "./styles/Checkbox.module.scss";
+import { useHydrated } from "@/hooks/useHydrated";
 
 interface CheckboxProps {
   id: string;
   checked: boolean;
   onChange: (checked: boolean) => void;
   label?: string;
+  transparent?: boolean;
 }
 
-export const Checkbox = ({ id, checked, onChange, label }: CheckboxProps) => {
+export const Checkbox = ({
+  id,
+  checked,
+  onChange,
+  label,
+  transparent = false,
+}: CheckboxProps) => {
+  const isHydrated = useHydrated();
+
   return (
     <div className={styles.checkboxContainer}>
-      <input
-        type="checkbox"
-        id={id}
-        checked={checked}
-        onChange={(e) => onChange(e.target.checked)}
-        className={styles.checkbox}
-      />
+      <div
+        className={classNames(styles.checkboxWrapper, {
+          [styles.transparent]: transparent,
+        })}
+        onClick={() => onChange(!checked)}
+      >
+        {isHydrated && checked && (
+          <div
+            className={classNames(styles.checkboxIcon, {
+              [styles.transparentIcon]: transparent,
+            })}
+          >
+            ✓
+          </div>
+        )}
+      </div>
       {label && (
-        <label htmlFor={id} className={styles.checkboxLabel}>
+        <label
+          htmlFor={id}
+          className={styles.checkboxLabel}
+          onClick={() => onChange(!checked)}
+        >
           {label}
         </label>
       )}