+++ src/hooks/useHydrated.tsx
@@ -1,0 +1,13 @@
+import { useState, useEffect } from "react";
+
+// Hook made to fix hydration errors in Chrome
+
+export const useHydrated = () => {
+  const [hydrated, setHydrated] = useState(false);
+
+  useEffect(() => {
+    setHydrated(true);
+  }, []);
+
+  return hydrated;
+};